Fixed heaps of warnings
- Removed some unnecessary virtual functions/overrides, since the generators no longer share the same noise octave functions
- Added some default values to certain undefined block behaviors
- Fixed some shadowing
- Fixed some sequence-uncertain variable modification
- Fixed wrong type for override/virtual

+++ src/bpp_shared/inventory/inventory.h
@@ -27,7 +27,7 @@ struct Inventory {
         return int(slots.size());
     }
 
-    virtual int getNetworkSlotId(int slot) const {
+    virtual int8_t getNetworkSlotId(int slot) const {
         if (slot < 0 || slot >= int(slots.size())) return -1;
         return slot;
     }
